requirements.txt
opencv-python
pyyaml
scipy
tensorboard
termcolor
timm==0.4.12
yacs